reorganize tests.
